+++ dptb/negf/hamiltonian.py
@@ -3,15 +3,13 @@ import torch
 from dptb.negf.areshkin_pole_sum import pole_maker
 from dptb.negf.RGF import recursive_gf
 from dptb.negf.surface_green import selfEnergy
-from dptb.negf.utils import quad, gauss_xw
+from dptb.negf.negf_utils import quad, gauss_xw,update_kmap,leggauss
 from dptb.negf.ozaki_res_cal import Ozaki_residues
 from dptb.negf.areshkin_pole_sum import pole_maker
 from ase.io import read
 from dptb.negf.poisson import density2Potential, getImg
 from dptb.negf.SCF import _SCF
 from dptb.utils.constants import *
-from dptb.negf.utils import update_kmap
-from dptb.negf.utils import leggauss
 import logging
 import os
 import torch.optim as optim
